requirements.txt
broadscope_bailian==1.3.1
dashscope==1.15.0
gradio==3.40.0
langchain_community==0.2.4
modelscope==1.12.0
Requests==2.32.3
tavily_python==0.3.3
torch==1.13.1
transformers==4.39.3